+++ project/tests/posts.py
@@ -9,7 +9,8 @@ from fastapi.testclient import TestClient
 import project.src.config.db.database as _database
 from project.src.app.main import app
 from project.src.app.routes.shared_constants_and_methods import (
-    SUCCESSFUL_DELETION_MESSAGE_KEY, SUCCESSFUL_DELETION_MESSAGE_VALUE_FOR_POST)
+    SUCCESSFUL_DELETION_MESSAGE_KEY, SUCCESSFUL_DELETION_MESSAGE_VALUE_FOR_POST, REQUEST_IS_OK_STATUS_CODE,
+    POST_ENTITY_BAD_TYPING_ERROR_STATUS_CODE, FORBIDDEN_REQUEST_STATUS_CODE, get_forbidden_request_detail_message)
 from project.src.app.routes.posts import get_db, posts_router
 
 load_dotenv()
@@ -44,16 +45,26 @@ test_post_owner_id = 1
 test_post2_id = ""
 
 
-def test_fetch_posts():
+def test_fetch_posts_should_succeed():
     response = posts_client.get(f"{posts_router.prefix}/")
-    assert response.status_code == 200, response.text
+    assert response.status_code == REQUEST_IS_OK_STATUS_CODE, response.text
+    data = response.json()
+    assert data == [], f"Should be [] because there is no posts yet!"
+
+    response = posts_client.get(f"{posts_router.prefix}/?owners=1&owners=3&skip=2")
+    assert response.status_code == REQUEST_IS_OK_STATUS_CODE, response.text
     data = response.json()
     assert data == [], f"Should be [] because there is no posts yet!"
 
 
+def test_fetch_posts_should_fail():
+    response = posts_client.get(f"{posts_router.prefix}/?owners=mike")
+    assert response.status_code == POST_ENTITY_BAD_TYPING_ERROR_STATUS_CODE, response.text
+
+
 def test_fetch_latest_posts():
     response = posts_client.get(f"{posts_router.prefix}/latest/")
-    assert response.status_code == 200, response.text
+    assert response.status_code == REQUEST_IS_OK_STATUS_CODE, response.text
     data = response.json()
     assert data == [], f"Should be [] because there is no posts yet!"
 
@@ -71,7 +82,7 @@ def test_create_post_should_succeed():
         },
     )
 
-    assert response.status_code == 200, response.text
+    assert response.status_code == REQUEST_IS_OK_STATUS_CODE, response.text
     data = response.json()
     assert data["caption"] == test_post_caption, f"Should be '{test_post_caption}'!"
     assert data["owner_id"] == test_post_owner_id, f"Should be '{test_post_owner_id}'!"
@@ -86,9 +97,24 @@ def test_create_post_should_succeed():
     test_post_id = data["id"]
 
 
+def test_create_post_should_fail():
+    response = posts_client.post(
+        f"{posts_router.prefix}/new",
+        json={
+            "image": "image test",
+            "caption": 45,
+            "tags": "lemon",
+            "published": False,
+            "owner_id": "mole",
+        },
+    )
+
+    assert response.status_code == POST_ENTITY_BAD_TYPING_ERROR_STATUS_CODE, response.text
+
+
 def test_get_post_should_succeed():
     response = posts_client.get(f"{posts_router.prefix}/{test_post_id}")
-    assert response.status_code == 200, response.text
+    assert response.status_code == REQUEST_IS_OK_STATUS_CODE, response.text
     data = response.json()
     assert data["caption"] == test_post_caption, f"Should be '{test_post_caption}'!"
     assert data["owner_id"] == test_post_owner_id, f"Should be '{test_post_owner_id}'!"
@@ -104,10 +130,17 @@ def test_get_post_should_fail():
     assert response.json() == {"detail": f"The post with id: {post_id} cannot be found!"}
 
 
-# I stopped on line 117 (routes/posts)
+def test_delete_post_should_fail():
+    user_id = 52
+    # Delete the post
+    response = posts_client.delete(f"{posts_router.prefix}/delete/{test_post_id}?user_id={user_id}")
+    assert response.status_code == FORBIDDEN_REQUEST_STATUS_CODE, response.text
+    assert response.json() == {"detail": get_forbidden_request_detail_message()}
+
+
 def test_delete_post_should_succeed():
     # Delete the post
-    response = posts_client.delete(f"{posts_router.prefix}/delete/{test_post_id}")
+    response = posts_client.delete(f"{posts_router.prefix}/delete/{test_post_id}?user_id={test_post_owner_id}")
     data = response.json()
     assert f"{ SUCCESSFUL_DELETION_MESSAGE_KEY }" in data
     assert data[f"{SUCCESSFUL_DELETION_MESSAGE_KEY}"] == f"{SUCCESSFUL_DELETION_MESSAGE_VALUE_FOR_POST}", \
